Only delete cache on Github Action

+++ src/__main__.py
@@ -123,6 +123,10 @@ class CloudflareManager:
             utils.save_cache(self.cache)
 
 
+def is_running_in_github_actions():
+    return os.getenv('GITHUB_ACTIONS') == 'true'
+
+
 def main():
     parser = argparse.ArgumentParser(description="Cloudflare Manager Script")
     parser.add_argument("action", choices=["run", "leave"], help="Choose action: run or leave")
@@ -131,11 +135,13 @@ def main():
     
     if args.action == "run":
         cloudflare_manager.update_resources()
-        utils.delete_cache()
+        if is_running_in_github_actions():
+            utils.delete_cache()
     elif args.action == "leave":
         cloudflare_manager.delete_resources()
     else:
         error("Invalid action. Please choose either 'run' or 'leave'.")
+
 
 if __name__ == "__main__":
     main()
